+++ receiver/statsdreceiver/internal/protocol/statsd_parser_test.go
